+++ src/common/wif.h
@@ -3,15 +3,7 @@
 #include <stddef.h>
 #include <stdint.h>
 #include <stdbool.h>
-
-// [x-coordinate (32)][y-coordinate (32)]
-#define PUBKEY_UNCOMPRESSED_LEN 64
-
-// [prefix (1)][x-coordinate (32)]
-#define PUBKEY_COMPRESSED_LEN 33
-
-// [wif (53)][\0]
-#define PUBKEY_WIF_STR_LEN 54
+#include "constants.h"
 
 /**
  * Convert raw public key into WIF representation supported by Hive backend