+++ src/crypto.c
@@ -2,7 +2,7 @@
  *   Ledger App Hive.
  *   (c) 2020 Ledger SAS.
  *   Modification based on (c) 2020 Andrew Chaney
- *   Modifications (c) 2021 Bartłomiej (@engrave) Górnicki
+ *   Modifications (c) 2021, 2022 Bartłomiej (@engrave) Górnicki
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -33,7 +33,10 @@
 uint8_t const SECP256K1_N[32] = {0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe,
                                  0xba, 0xae, 0xdc, 0xe6, 0xaf, 0x48, 0xa0, 0x3b, 0xbf, 0xd2, 0x5e, 0x8c, 0xd0, 0x36, 0x41, 0x41};
 
-int crypto_derive_private_key(cx_ecfp_private_key_t *private_key, uint8_t chain_code[static 32], const uint32_t *bip32_path, uint8_t bip32_path_len) {
+int crypto_derive_private_key(cx_ecfp_private_key_t *private_key,
+                              uint8_t chain_code[static CHAINCODE_LEN],
+                              const uint32_t *bip32_path,
+                              uint8_t bip32_path_len) {
     uint8_t raw_private_key[32] = {0};
 
     BEGIN_TRY {
@@ -55,7 +58,7 @@ int crypto_derive_private_key(cx_ecfp_private_key_t *private_key, uint8_t chain_
     return 0;
 }
 
-int crypto_init_public_key(cx_ecfp_private_key_t *private_key, cx_ecfp_public_key_t *public_key, uint8_t raw_public_key[static 64]) {
+int crypto_init_public_key(cx_ecfp_private_key_t *private_key, cx_ecfp_public_key_t *public_key, uint8_t raw_public_key[static PUBKEY_UNCOMPRESSED_LEN]) {
     // generate corresponding public key
     cx_ecfp_generate_pair(CX_CURVE_256K1, public_key, private_key, 1);
 
@@ -64,7 +67,7 @@ int crypto_init_public_key(cx_ecfp_private_key_t *private_key, cx_ecfp_public_ke
     return 0;
 }
 
-bool crypto_sign_message() {
+bool crypto_sign_digest(const uint8_t digest[static DIGEST_LEN], uint8_t signature[static SIGNATURE_LEN]) {
     cx_ecfp_private_key_t private_key = {0};
     uint8_t chain_code[CHAINCODE_LEN] = {0};
     uint8_t der_signature[MAX_DER_SIG_LEN] = {0};
@@ -74,10 +77,7 @@ bool crypto_sign_message() {
     int32_t counter = 0;
     bool is_valid = true;
 
-    // store hash (take 0 bytes from current hash and copy it to the output buffer)
-    cx_hash((cx_hash_t *) &G_context.tx_info.sha, CX_LAST, G_context.tx_info.m_hash, 0, G_context.tx_info.m_hash, sizeof(G_context.tx_info.m_hash));
-
-    PRINTF("Digest: %.*H\n", DIGEST_LEN, G_context.tx_info.m_hash);
+    PRINTF("Digest: %.*H\n", DIGEST_LEN, digest);
 
     // derive private key according to BIP32 path
     crypto_derive_private_key(&private_key, chain_code, G_context.bip32_path, G_context.bip32_path_len);
@@ -89,30 +89,23 @@ bool crypto_sign_message() {
              * digest before hashing. This results in a new deterministic k each round which will result in either a canonical or non-canonical signature. */
             while (true) {
                 if (counter == 0) {
-                    rng_rfc6979(der_signature, G_context.tx_info.m_hash, private_key.d, private_key.d_len, SECP256K1_N, ARRAYLEN(SECP256K1_N), V, K);
+                    rng_rfc6979(der_signature, (uint8_t *) digest, private_key.d, private_key.d_len, SECP256K1_N, ARRAYLEN(SECP256K1_N), V, K);
                 } else {
-                    rng_rfc6979(der_signature, G_context.tx_info.m_hash, NULL, 0, SECP256K1_N, ARRAYLEN(SECP256K1_N), V, K);
+                    rng_rfc6979(der_signature, (uint8_t *) digest, NULL, 0, SECP256K1_N, ARRAYLEN(SECP256K1_N), V, K);
                 }
 
-                cx_ecdsa_sign(&private_key,
-                              CX_NO_CANONICAL | CX_RND_PROVIDED | CX_LAST,
-                              CX_SHA256,
-                              G_context.tx_info.m_hash,
-                              DIGEST_LEN,
-                              der_signature,
-                              MAX_DER_SIG_LEN,
-                              &info);
+                cx_ecdsa_sign(&private_key, CX_NO_CANONICAL | CX_RND_PROVIDED | CX_LAST, CX_SHA256, digest, DIGEST_LEN, der_signature, MAX_DER_SIG_LEN, &info);
 
                 if ((info & CX_ECCINFO_PARITY_ODD) != 0) {
                     der_signature[0] |= 0x01;
                 }
 
-                if (!signature_from_der(der_signature, G_context.tx_info.signature, MEMBER_SIZE(transaction_ctx_t, signature))) {
+                if (!signature_from_der(der_signature, signature, MEMBER_SIZE(transaction_ctx_t, signature))) {
                     is_valid = false;
                     break;
                 }
 
-                if (signature_check_canonical(G_context.tx_info.signature + 1)) {
+                if (signature_check_canonical(signature + 1)) {
                     break;
                 } else {
                     counter++;