+++ src/helper/send_response.h
@@ -37,4 +37,4 @@ int helper_send_response_pubkey(void);
  * @return zero or positive integer if success, -1 otherwise.
  *
  */
-int helper_send_response_sig(void);
+int helper_send_response_sig(const uint8_t *signature, size_t sig_len);