+++ src/sw.h
@@ -48,3 +48,15 @@
  * Status word for signature fail.
  */
 #define SW_SIGNATURE_FAIL 0xB005
+/**
+ * Status word for trying to sign hash with policy disabled in settings.
+ */
+#define SW_HASH_SIGNING_DISABLED 0xB006
+/**
+ * Status word invalid length of input dat.
+ */
+#define SW_WRONG_HASH_LENGTH 0xB007
+/**
+ * Status word for hash parsing fail.
+ */
+#define SW_HASH_PARSING_FAIL 0xB008