+++ src/transaction/transaction_parse.h
@@ -15,3 +15,12 @@
  *
  */
 parser_status_e transaction_parse(buffer_t *buf);
+
+/**
+ * @brief Parse incoming path and digest
+ *
+ * @param buf
+ *  Pointer to buffer with path and digest
+ * @return PARSING_OK if success, error status otherwise.
+ */
+parser_status_e hash_parse(buffer_t *buf);