+++ src/ui/screens/review_hash.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdbool.h>
+
 #include "os.h"
 #include "ux.h"
 #include "glyphs.h"
@@ -8,18 +10,22 @@
 #include "globals.h"
 #include "io.h"
 #include "sw.h"
-#include "ui/action/validate.h"
 #include "common/bip32.h"
 #include "common/format.h"
 #include "common/wif.h"
 #include "common/macros.h"
-#include "transaction/decoders.h"
-#include "transaction/transaction_parse.h"
+#include "ui/action/validate.h"
 
 /**
- * Display address on the device and ask confirmation to export.
+ * Display hash string on the device and ask confirmation before signing
  *
  * @return 0 if success, negative integer otherwise.
  *
  */
-int ui_display_public_key(void);
+int ui_display_hash(void);
+
+/**
+ * Initialize "Signing hash" display when digest got accepted
+ *
+ */
+void ui_display_signing_hash_message(void);