+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-04-13
+
+### Feature
+
+- Settings menu to enable/disable hash signing
+- Hash signing command
+
 ## [1.0.1] - 2022-02-20
 
 ### Fixed
 
 - update_account uses authority type optional serializer
-
 
 ## [1.0.0] - 2021-08-29